+++ searchbase/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 3.0.0
+
+- Add support for [KeysToSubscribe] enum.
+- Add support for [ChangesController] class to read the state changes in subscriber method.
+- Breaking Change
+  - `subscribeToStateChanges()` now accepts array of enums ([KeysToSubscribe]) for properties to subscribe, instead of hard-coded strings.
+    Also, the subscription callback function receives a [ChangesController] class argument instead of [Changes] type. - **before**
+    `dart widgetInstance.subscribeToStateChanges((Changes change){ print('${changes['results']!.next}'); }, ['results']); ` - **after**
+    `dart widgetInstance.subscribeToStateChanges((ChangesController change){ print('${changes.Results!.next}'); }, [KeysToSubscribe.Results]); `
+    > Refer to docs to see the [KeysToSubscribe] enum defined under _types.dart_
+
 ## 2.2.4
 
 - Fix callbacks for change events